Reorganize tx validator to reduce go sdk memory usage on mobile
